refactor: Change Auth to ClientProvider
