--- src/auth/index.ts
@@ -1,4 +1,0 @@
-export { NtlmAuth } from "./ntlm";
-export { OAuth } from "./oauth";
-export { ConnectionStringAuth } from "./connection-string-auth";
-export { Auth } from "./auth";
