+++ src/client/auth/oauth/index.ts
@@ -1,3 +1,3 @@
-export { OAuth } from "./oauth";
+export { OAuthClientProvider } from "./oauth-client-provider";
 export { discoverAuthority } from "./authority";
 export {  DeviceCodeResponse, OAuthConfig as OAuth2Config, OAuth2Credentials, PersistenceOptions } from "./oauth-configuration";
