+++ src/client/auth/oauth/msal/extensions/Dpapi.ts
