+++ src/client/auth/oauth/msal/extensions/error/PersistenceError.ts
