+++ src/client/auth/oauth/msal/extensions/index.ts
