+++ src/client/auth/oauth/msal/extensions/lock/CrossPlatformLock.ts
