+++ src/client/auth/oauth/msal/extensions/lock/CrossPlatformLockOptions.ts
