+++ src/client/auth/oauth/msal/extensions/persistence/BasePersistence.ts
