+++ src/client/auth/oauth/msal/extensions/persistence/DataProtectionScope.ts
