+++ src/client/auth/oauth/msal/extensions/persistence/FilePersistenceWithDataProtection.ts
