+++ src/client/auth/oauth/msal/extensions/persistence/IPersistence.ts
