+++ src/client/auth/oauth/msal/extensions/persistence/IPersistenceConfiguration.ts
