+++ src/client/auth/oauth/msal/extensions/persistence/PersistenceCachePlugin.ts
