+++ src/client/auth/oauth/msal/extensions/utils/Constants.ts
