+++ src/client/auth/oauth/msal/extensions/utils/Environment.ts
