+++ src/client/auth/oauth/msal/token/application.ts
