+++ src/index.ts
@@ -17,5 +17,4 @@ export * from "./dataverse-client-options";
 
 export { DataverseClient, EntityCollection } from "./dataverse-client";
 
-export * from "./connection-string";
-export * from "./auth";
+export * from "./client";
